requirements.txt
django
pytz==2017.2
djangorestframework==3.5.3